REFACTOR: move tests to separate test case
reduce the number of tests run by 63
