+++ tests/base_photo_test_case.py
@@ -2,7 +2,6 @@ import hashlib
 import os
 from shutil import rmtree, copyfile
 
-from family_foto.models import db
 from family_foto.models.file import File
 from family_foto.models.photo import Photo
 from tests.base_test_case import BaseTestCase
@@ -32,11 +31,3 @@ class BasePhotoTestCase(BaseTestCase):
             if not os.path.exists(copied_path):
                 raise FileNotFoundError(f'{copied_path} does not exists.')
             self.photo = Photo(filename='example.jpg', hash=file_hash)
-
-    def test_commit(self):
-        """Tests committing the file works"""
-        db.session.add(self.photo)
-        db.session.commit()
-
-        photo = Photo.query.get(self.photo.id)
-        self.assertEqual(self.photo, photo)
