+++ tests/base_test_case.py
@@ -4,8 +4,9 @@ from shutil import rmtree
 from flask_testing import TestCase
 
 import family_foto
-from family_foto import add_user, Role
+from family_foto import add_user
 from family_foto.models import db
+from family_foto.models.role import Role
 from family_foto.models.user import User
 from family_foto.models.user_settings import UserSettings
 
@@ -44,13 +45,3 @@ class BaseTestCase(TestCase):
         db.session.close()
         db.drop_all()
         rmtree(self.app.instance_path)
-
-    def test_setup(self):
-        """
-        Tests the test database is clean before each test.
-        """
-        user = User.query.filter_by(username='marcel').first()
-        settings = UserSettings.query.filter_by(user_id=user.id).first()
-
-        self.assertIsNotNone(user.settings)
-        self.assertEqual(user, settings.user)
